Finished wrapping the robot input functions.
Switched to using a location for where to position the mouse. This means that either a selector or a position object can be passed in to define where to move the mouse to.

Still need to determine how to start/stop the robot so that it will run correctly on each test case. May need to think of a different approach. The way doh works it to have the robot run once after all of the tests have been specified. Not sure if we can do this in the qunit environment.